requirements.txt
numpy
tensorflow
opencv-python
picamera2
pyserial
